backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
python-multipart==0.0.20
pydantic==2.9.2
python-dotenv==1.0.1
httpx==0.27.2
elasticsearch==8.15.1
numpy==2.1.3
scipy==1.14.1
orjson==3.10.11
jinja2==3.1.4
tenacity==9.0.0
websockets==13.1
pydantic-settings==2.6.1
pypdf==6.1.1
python-docx==1.1.2
python-pptx==1.0.2
